pyproject.toml
[project]
name = "stock-market-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.2.5",
    "fpdf>=1.7.2",
    "gtts>=2.5.4",
    "httpx>=0.28.1",
    "matplotlib>=3.10.1",
    "mcp[cli]>=1.6.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "yfinance>=0.2.56",
]
